Compose Mono/Flux for better efficiency
Instead of relying on `parallelStream` to make concurrent remote
blocking calls, we can leverage `Mono` and `Flux` composition
operators.

We can then `Mono.zip` all remote calls into one, reactive
pipeline. We'll still need to block on it before putting it in the
`Model` map given to the view template.